+++ src/main/java/jeziel/graficadora/Modelos/Punto3D.java
@@ -2,41 +2,41 @@ package jeziel.graficadora.Modelos;
 
 public class Punto3D extends Punto<Punto3D, Vector3D> {
 
-    public Punto3D(float ordenadaX, float ordenadaY, float ordenadaZ) {
+    public Punto3D(double ordenadaX, double ordenadaY, double ordenadaZ) {
         this.ordenadaX = ordenadaX;
         this.ordenadaY = ordenadaY;
         this.ordenadaZ = ordenadaZ;
     }
 
-    public float getOrdenadaX() {
+    public double getOrdenadaX() {
         return ordenadaX;
     }
 
-    public void setOrdenadaX(float ordenadaX) {
+    public void setOrdenadaX(double ordenadaX) {
         this.ordenadaX = ordenadaX;
     }
 
-    public float getOrdenadaY() {
+    public double getOrdenadaY() {
         return ordenadaY;
     }
 
-    public void setOrdenadaY(float ordenadaY) {
+    public void setOrdenadaY(double ordenadaY) {
         this.ordenadaY = ordenadaY;
     }
 
-    public float getOrdenadaZ() {
+    public double getOrdenadaZ() {
         return ordenadaZ;
     }
 
-    public void setOrdenadaZ(float ordenadaZ) {
+    public void setOrdenadaZ(double ordenadaZ) {
         this.ordenadaZ = ordenadaZ;
     }
 
     @Override
-    public float obtenerDistanciaPunto(Punto3D p) {
-        float dx = this.ordenadaX - p.ordenadaX;
-        float dy = this.ordenadaY - p.ordenadaY;
-        float dz = this.ordenadaZ - p.ordenadaZ;
+    public double obtenerDistanciaPunto(Punto3D p) {
+        double dx = this.ordenadaX - p.ordenadaX;
+        double dy = this.ordenadaY - p.ordenadaY;
+        double dz = this.ordenadaZ - p.ordenadaZ;
         return (float) Math.sqrt(dx*dx + dy*dy + dz*dz);
     }
