+++ src/main/java/jeziel/graficadora/Modelos/Vector3D.java
@@ -2,13 +2,13 @@ package jeziel.graficadora.Modelos;
 
 public class Vector3D extends Vector<Vector3D>{
 
-    public Vector3D(float vectorX,float vectorY,float vectorZ) {
+    public Vector3D(double vectorX,double vectorY,double vectorZ) {
         this.vectorX = vectorX;
         this.vectorY = vectorY;
         this.vectorZ = vectorZ;
     }
 
-    public float getVectorX() {
+    public double getVectorX() {
         return vectorX;
     }
 
@@ -16,7 +16,7 @@ public class Vector3D extends Vector<Vector3D>{
         this.vectorX = vectorX;
     }
 
-    public float getVectorY() {
+    public double getVectorY() {
         return vectorY;
     }
 
@@ -24,7 +24,7 @@ public class Vector3D extends Vector<Vector3D>{
         this.vectorY = vectorY;
     }
 
-    public float getVectorZ() {
+    public double getVectorZ() {
         return vectorZ;
     }
 
@@ -33,7 +33,7 @@ public class Vector3D extends Vector<Vector3D>{
     }
 
     @Override
-    public float obtenerMagnitud(){
+    public double obtenerMagnitud(){
         return (float) Math.sqrt(vectorX*vectorX+vectorY*vectorY+vectorZ*vectorZ);
     }
 
@@ -48,18 +48,18 @@ public class Vector3D extends Vector<Vector3D>{
     }
 
     @Override
-    public Vector3D multEscalar(float k){
+    public Vector3D multEscalar(double k){
         return new Vector3D(this.vectorX*k,this.vectorY*k,this.vectorZ*k);
     }
 
     @Override
-    public float obtenerProductoPunto(Vector3D v){
+    public double obtenerProductoPunto(Vector3D v){
         return (this.vectorX*v.vectorX)+(this.vectorY*v.vectorY)+(this.vectorZ*v.vectorZ);
     }
 
     @Override
     public Vector3D normalizar(){
-        float mag=obtenerMagnitud();
+        double mag=obtenerMagnitud();
         if (mag==0){
             throw new ArithmeticException("No se puede normalizar vector nulo");
         }
